Set the same output for different programs

+++ autoTransport/Bus.h
@@ -6,5 +6,5 @@ struct Bus
 	type mKey;
 	int tPower;
 	short int mData;
-	double  fuelConsumption;
+	float fuelConsumption;
 };
