+++ autoTransport/Car.h
@@ -6,5 +6,5 @@ struct Car
 	type mKey;
 	int tPower;
 	short int mData;
-	double  fuelConsumption;
+	float fuelConsumption;
 };
