+++ autoTransport/Truck.h
@@ -6,5 +6,5 @@ struct Truck
 	type mKey;
 	int tPower;
 	int mData;
-	double fuelConsumption;
+	float fuelConsumption;
 };
